+++ sdk/webpubsub/azure-messaging-webpubsub-client/src/test/java/com/azure/messaging/webpubsub/client/ClientTests.java
@@ -252,7 +252,7 @@ public class ClientTests extends TestBase {
     public void testBothCredential() {
         Assertions.assertThrows(IllegalStateException.class, () -> {
             WebPubSubClient client = new WebPubSubClientBuilder()
-                .credential(new WebPubSubClientCredential(Mono.just("mock")))
+                .credential(new WebPubSubClientCredential(() ->"mock"))
                 .clientAccessUrl("mock")
                 .buildClient();
         });